Major refactoring and upgrades
basic.py is now atlas.py; it is no longer intended to be ran from the terminal and instead it needs to be imported by Python scripts and used as a library. Settings are now stored as objects of class Settings() rather than being fixed for all runs. Added scripts download.com and compile.com that download all necessary files (line databases, Fortran scripts etc) from external sources and compile executables. Added test.py to test if the setup is complete. This is the first commit in a series: more to follow