Check ODF temp range, allow disabling molecules
Added a check at the end of each ATLAS run to make sure the temperatures throughout the atmosphere do not exceed the ODF range. Allowed disabling molecular treatment in ATLAS which may be useful at high temperatures where no molecules are known to exist. When molecules are disabled, ATLAS solves the Saha equation directly allowing for a more reliable result (otherwise an approximate iterative method is employed). Replaced the numerical finder of the dB/dT maximum with an analytic one for performance