Fixed ATLAS-to-standard abundance conversion
So far we've assumed that ATLAS metal abundances are given with respect to N(H)+N(He) as was originally stated in the ATLAS Cookbook (no longer available). This is not actually true. All abundances are given as fractions of the total nuclear number density. At solar or subsolar metallicities, the difference is negligible; however, it may become more prominent at supersolar metallicities. This commit fixes it.