Major updates. See commit description for details. Reinstall required
(1) De-escalate errors thrown when the ODF or the wavelength grid is exceeded in ATLAS to warnings. Most of the time it does not make a noticeable difference and we don't need to crash the code. (2) Added the option to adjust the C12/C13 ratio in molecular opacities. (3) Replaced the broken CH linelist with the corrected version. (4) Replaced the MgH line list with a newer version from ExoMol. (4) Provided 4 alternative atomic line lists to the default Kurucz one, and set the one that matches the spectrum of Arcturus best as the new default. (5) Removed predicted lines from SYNTHE, since the latest versions of the atomic and molecular line lists supercede them. (6) Minor bug fixes and documentation updates.