Prevented ATLAS from getting stuck due to bad chemical equilibrium
Sometimes ATLAS gets stuck indefinitely when it cannot find chemical equilibrium. I patched the codebase to limit the number of chemical solver iterations to 20000, and to report in the output file when the chemical solver fails. Also forced the convergence of models where the last iteration did not have a good chemical solution to be 88888.888