Major update. Clean and recompile required. See commit description
Changed the mechanism for automatic iterations in ATLAS. Instead of running each batch as a separate instance of ATLAS, run all iterations together, and have ATLAS stop itself once the convergence goal is reached. niter now sets the maximum number of iterations. Take the best iteration in the end. Reformat the output in output_main.out to remove all ambiguities