Major changes to the installation procedure, see commit description
(1) Added checks to both download.com and compile.com that ensure that all of the required commands are available in the shell before running. (2) Added a prompt to the download script, asking the user to download the full grid of restart models from Google Drive. (3) Updated reduce_tio.py to use sys.argv instead of user prompts. (4) Added a call to reduce_tio.py in the compile script to reduce the TiO linelist by default. (5) Updated test.py to expect the hashsum of the reduced TiO list and to check the hashsum of the full grid of restarts if it is available